[giac] fsolve change for NSolve
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@25365 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/giac/modpoly.cc
@@ -2038,7 +2038,7 @@ namespace giac {
 
   bool convertfromdouble(const vector<double> & A,modpoly & a,double M){
     a.clear(); a.reserve(A.size());
-    int m(M);
+    int m( (int)M);
     vector<double>::const_iterator it=A.begin(),itend=A.end();
     for (;it!=itend;++it){
       double d=*it;
